rename 'acle' to 'arm_shared'
