+++ crates/core_arch/src/aarch64/mod.rs
@@ -21,7 +21,7 @@ pub use self::crc::*;
 mod prefetch;
 pub use self::prefetch::*;
 
-pub use super::acle::*;
+pub use super::arm_shared::*;
 
 #[cfg(test)]
 use stdarch_test::assert_instr;
