+++ crates/core_arch/src/arm_shared/barrier/common.rs
