+++ crates/core_arch/src/arm_shared/barrier/cp15.rs
