+++ crates/core_arch/src/arm_shared/crc.rs
