+++ crates/core_arch/src/arm_shared/ex.rs
