+++ crates/core_arch/src/arm_shared/hints.rs
