+++ crates/core_arch/src/arm_shared/neon/generated.rs
