+++ crates/core_arch/src/arm_shared/test_support.rs
