+++ crates/stdarch-gen/src/main.rs
@@ -2253,7 +2253,7 @@ mod test {
 
     let arm_out_path: PathBuf = PathBuf::from(env::var("OUT_DIR").unwrap())
         .join("src")
-        .join("acle")
+        .join("arm_shared")
         .join("neon");
     std::fs::create_dir_all(&arm_out_path)?;
